+++ Procfile
@@ -1,1 +1,1 @@
-web: sh setup.sh && streamlit run myapp2.py
+web: sh setup.sh && streamlit run my-app2.py
